1 - bug fix
when the relation class is not in the same package than the classes that play some role in the relation, the generated code did not compile.

2 - new feature

possibility of declaring new ValueTypes that can be used to declare slots in classes.  The syntax for this is the following:

   valueType <typeName> [ as <aliasName> ] ;

That is, we can simply say something like:

   valueType java.math.BigInteger;

and, after that, declare entities like this:

   class MyEntity {
       String someTxt;
       java.math.BigInteger myBigIntSlot;
   }

The extended syntax allows the use of an alias to the type.
With this extended syntax, the example above becomes:

   valueType java.math.BigInteger as Bignum;

   class MyEntity {
       String someTxt;
       Bignum myBigIntSlot;
   }

This "alias syntax" was added also to the external declaration.
Now it is possible to write something like this:

   external class my.big.external.entity.SomeName as ShortName;


****IMPORTANT NOTE****:
This new declaration should be used only for types that have
value-type semantics: types with immutable instances.
The inclusion of such a declaration on a DML file is a promise made by
the programmer to the DML compiler that the elements of the given
class are immutable and serializable.  Later, they will have to
participate on some kind of protocol that allows them to be written
and read back from a database.