New version of the DML compiler with rudimentary support for enum value-types.
It is now possible to declare a value-type that corresponds to an Enum
by replacing the "valueType" keyword with the new "enum" keyword.
Value-types declared in this way are handled specially by the Fenix
framework: it is assumed that slots of these types are stored in
VARCHAR columns on the database, and the conversion to and from the
database is automatically supported.  Thus, it is no longer needed to
create FieldConversion subclasses to handle enums.

In these set of changes, all the converter classes for enums where
deleted.