Bug fixed in the GEP section.
In the reports by degree type, the teachers list (from GIAF and from APLICA) is revised. Two columns were not correct ("Academic Degree" and "Degree name or area") in both cases.
The "Academic Degree" is supposed to be the best qualification type, and the "Degree name or area" is supposed to be what was on the "Academic Degree" column before.