Upgrade to the latest fenix-framework JAR, which uses only OIDs as foreign keys.
WARNING: Before you upgrade to this version of the fenix application,
make sure that your database has all the OIDs correctly defined.
Otherwise, upgrading to this version of the fenix-framework will not
work with your current database.

Use the pt.ist.fenixframework.pstm.repository.CheckOids program to
check if the OIDs are OK.