Fixed search space by person
The interface is now re-enabled and adapted to the new fenixedu-spaces.
The layoyt was fixed as the elements, instead of being displayed inline, were stacked.
In the resulting page we now check if the blueprint exists before printing it.
Also in the resulting page, the occupants name are displayed, with the link to their webpages if these are available