Made further changes to group names to avoid multiple Groups sharing the same name.
Added more info to the execution course part of the Teacher, TeacherResponsibleOfExecutionCourse and Student Group names.
Added info on whether a UnitGroup includes subunits to it's name. Also corrected a bug with the relation type part of
the UnitGroup's name. Corrected a NullPointerException on calling AlumniGroup.getPresentationName on an instance with a null degree.